--- packages/storybook/babel.config.json
@@ -1,5 +1,0 @@
-{
-  "sourceType": "unambiguous",
-  "presets": [["@babel/preset-env"], "@babel/preset-typescript"],
-  "plugins": [["@babel/plugin-proposal-decorators", { "decoratorsBeforeExport": true }]]
-}
