--- packages/storybook/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@boiler/storybook",
-  "version": "0.0.1",
-  "description": "",
-  "scripts": {
-    "start": "sb dev -p 1337",
-    "build:static": "sb build",
-    "create-config:pa11y": "node --experimental-json-modules utils/create-pa11y-config.mjs",
-    "pretest:a11y": "npm run build:static && npm run create-config:pa11y",
-    "test:a11y": "pa11y-ci --config ./.pa11yci"
-  },
-  "author": "",
-  "license": "MIT",
-  "devDependencies": {
-    "@boiler/ui-library": "^0.0.1",
-    "@storybook/addon-actions": "^7.5.3",
-    "@storybook/addon-essentials": "^7.5.3",
-    "@storybook/cli": "^7.5.3",
-    "@storybook/web-components": "^7.5.3",
-    "@storybook/web-components-webpack5": "^7.5.3",
-    "@storybook/manager-api": "^7.5.3",
-    "@storybook/theming": "^7.5.3"
-  },
-  "dependencies": {
-    "@storybook/addons": "^7.5.3"
-  }
-}
