+++ packages/ui-library/.storybook/boilerTheme.js
@@ -1,5 +1,5 @@
 import { create } from '@storybook/theming/create';
-import B01LER from '../public/B01LER.png'
+import B01LER from './public/B01LER.png';
 
 export default create({
   base: 'dark',
