+++ packages/ui-library/.storybook/preview-head.html
@@ -1,7 +1,7 @@
 <style>
   @font-face {
     font-family: 'SourceSansPro';
-    src: url('../static/fonts/Source_Sans_Pro/SourceSansPro-Regular.ttf') format('truetype');
+    src: url('./static/fonts/Source_Sans_Pro/SourceSansPro-Regular.ttf') format('truetype');
   }
   body {
     font-family: 'SourceSansPro';
