+++ packages/ui-library/babel.config.json
@@ -1,0 +1,8 @@
+{
+  "sourceType": "unambiguous",
+  "presets": [["@babel/preset-env"], "@babel/preset-typescript"],
+  "plugins": [
+    ["@babel/plugin-proposal-decorators", { "decoratorsBeforeExport": true }],
+    "@babel/plugin-transform-class-properties"
+  ]
+}
