+++ packages/ui-library/package.json
@@ -3,6 +3,8 @@
   "version": "0.0.1",
   "description": "",
   "scripts": {
+    "start:storybook": "sb dev -p 1337",
+    "build:storybook": "sb build",
     "build": "NODE_ENV=production webpack --config webpack.prod.js && npm run copy:static",
     "build:types": "rimraf types && tsc --emitDeclarationOnly",
     "lint": "npm run lint:eslint && npm run lint:stylelint",
